no usage of global.window check, because files are only run in a context in which they can run successfully
